Moved signal blocker and -handler to configuration module and corrected minor type errors
